Switched the ultimate control and entry to be separate controls
