+++ Library/PhantomLib/CustomControls/Enums/UltimateEntryImageButton.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace PhantomLib.CustomControls.Enums
+{
+    public enum UltimateEntryImageButton
+    {
+        None,
+        ClearContents,
+        Password
+    }
+}
