+++ Library/PhantomLib/CustomControls/Enums/UltimateEntryReturn.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace PhantomLib.CustomControls.Enums
+{
+    public enum UltimateEntryReturn
+    {
+        Next,
+        Done,
+        Search
+    }
+}
